feat(jdbc): rename VectorWise plugin to Actian Vector
