+++ plugin-jdbc-actianvector/src/main/java/io/kestra/plugin/jdbc/actianvector/ActianVectorConnectionInterface.java
@@ -1,0 +1,10 @@
+package io.kestra.plugin.jdbc.actianvector;
+
+import io.kestra.plugin.jdbc.JdbcConnectionInterface;
+
+public interface ActianVectorConnectionInterface extends JdbcConnectionInterface {
+    @Override
+    default String getScheme() {
+        return "jdbc:vectorwise";
+    }
+}
