+++ plugin-jdbc-actianvector/src/main/java/io/kestra/plugin/jdbc/actianvector/Queries.java
@@ -1,0 +1,80 @@
+package io.kestra.plugin.jdbc.actianvector;
+
+import com.ingres.jdbc.IngresDriver;
+import io.kestra.core.exceptions.IllegalVariableEvaluationException;
+import io.kestra.core.models.annotations.Example;
+import io.kestra.core.models.annotations.Metric;
+import io.kestra.core.models.annotations.Plugin;
+import io.kestra.core.models.executions.metrics.Counter;
+import io.kestra.core.runners.RunContext;
+import io.kestra.plugin.jdbc.AbstractCellConverter;
+import io.kestra.plugin.jdbc.AbstractJdbcQueries;
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+import lombok.experimental.SuperBuilder;
+
+import java.sql.DriverManager;
+import java.sql.SQLException;
+import java.time.ZoneId;
+
+@SuperBuilder
+@ToString
+@EqualsAndHashCode
+@Getter
+@NoArgsConstructor
+@Schema(
+    title = "Execute multiple SQL statements against Actian Vector",
+    description = "Runs multiple SQL statements against Actian Vector (formerly Vectorwise) columnar analytics database. Optimized for analytical workloads and high-performance queries. Supports parameterized queries, transactions (default enabled), and all fetch modes. Default fetchSize is 10,000 rows."
+)
+@Plugin(
+    aliases = "io.kestra.plugin.jdbc.vectorwise.Queries",
+    examples = {
+        @Example(
+            title = "Send SQL queries to an Actian Vector database and fetch a row as output.",
+            full = true,
+            code = """
+                   id: actianvector_queries
+                   namespace: company.team
+
+                   tasks:
+                     - id: queries
+                       type: io.kestra.plugin.jdbc.actianvector.Queries
+                       url: jdbc:vectorwise://url:port/base
+                       username: "{{ secret('VECTORWISE_USERNAME') }}"
+                       password: "{{ secret('VECTORWISE_PASSWORD') }}"
+                       sql: SELECT count(*) FROM employee; SELECT count(*) FROM laptop;
+                       fetchType: FETCH_ONE
+                   """
+        )
+    },
+    metrics = {
+        @Metric(
+            name = "fetch.size",
+            type = Counter.TYPE,
+            unit = "rows",
+            description = "The number of fetched rows."
+        )
+    }
+)
+public class Queries extends AbstractJdbcQueries implements ActianVectorConnectionInterface {
+    @Override
+    protected AbstractCellConverter getCellConverter(ZoneId zoneId) {
+        return new ActianVectorCellConverter(zoneId);
+    }
+
+    @Override
+    public void registerDriver() throws SQLException {
+        // only register the driver if not already exist to avoid a memory leak
+        if (DriverManager.drivers().noneMatch(IngresDriver.class::isInstance)) {
+            DriverManager.registerDriver(new IngresDriver());
+        }
+    }
+
+    @Override
+    protected Integer getFetchSize(RunContext runContext) throws IllegalVariableEvaluationException {
+        return runContext.render(this.fetchSize).as(Integer.class).orElse(10000);
+    }
+}
