+++ plugin-jdbc-actianvector/src/main/java/io/kestra/plugin/jdbc/actianvector/Trigger.java
@@ -1,0 +1,90 @@
+package io.kestra.plugin.jdbc.actianvector;
+
+import com.ingres.jdbc.IngresDriver;
+import io.kestra.core.models.annotations.Example;
+import io.kestra.core.models.annotations.Plugin;
+import io.kestra.core.models.property.Property;
+import io.kestra.core.runners.RunContext;
+import io.kestra.plugin.jdbc.AbstractJdbcQuery;
+import io.kestra.plugin.jdbc.AbstractJdbcTrigger;
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+import lombok.experimental.SuperBuilder;
+
+import java.sql.DriverManager;
+import java.sql.SQLException;
+
+@SuperBuilder
+@ToString
+@EqualsAndHashCode
+@Getter
+@NoArgsConstructor
+@Schema(
+    title = "Wait for query results on Actian Vector and trigger flow",
+    description = "Periodically polls an Actian Vector database by executing a SQL query at the specified interval (default 60 seconds). Triggers a downstream flow execution when the query returns one or more rows. Supports parameterized queries and afterSQL for marking processed rows. Use fetchType to control result handling."
+)
+@Plugin(
+    aliases = "io.kestra.plugin.jdbc.vectorwise.Trigger",
+    examples = {
+        @Example(
+            title = "Wait for a SQL query to return results, and then iterate through rows.",
+            full = true,
+            code = """
+                id: jdbc_trigger
+                namespace: company.team
+
+                tasks:
+                  - id: each
+                    type: io.kestra.plugin.core.flow.ForEach
+                    values: "{{ trigger.rows }}"
+                    tasks:
+                      - id: return
+                        type: io.kestra.plugin.core.debug.Return
+                        format: "{{ json(taskrun.value) }}"
+
+                triggers:
+                  - id: watch
+                    type: io.kestra.plugin.jdbc.actianvector.Trigger
+                    interval: "PT5M"
+                    url: jdbc:vectorwise://url:port/base
+                    username: "{{ secret('VECTORWISE_USERNAME') }}"
+                    password: "{{ secret('VECTORWISE_PASSWORD') }}"
+                    sql: "SELECT * FROM my_table"
+                    fetchType: FETCH
+                """
+        )
+    }
+)
+public class Trigger extends AbstractJdbcTrigger implements ActianVectorConnectionInterface {
+    @Override
+    protected AbstractJdbcQuery.Output runQuery(RunContext runContext) throws Exception {
+        var query = Query.builder()
+            .id(this.id)
+            .type(Query.class.getName())
+            .url(this.getUrl())
+            .username(this.getUsername())
+            .password(this.getPassword())
+            .timeZoneId(this.getTimeZoneId())
+            .sql(this.getSql())
+            .afterSQL(this.getAfterSQL())
+            .fetch(this.isFetch())
+            .store(this.isStore())
+            .fetchOne(this.isFetchOne())
+            .fetchType(Property.ofValue(this.renderFetchType(runContext)))
+            .additionalVars(this.additionalVars)
+            .parameters(this.getParameters())
+            .build();
+        return query.run(runContext);
+    }
+
+    @Override
+    public void registerDriver() throws SQLException {
+        // only register the driver if not already exist to avoid a memory leak
+        if (DriverManager.drivers().noneMatch(IngresDriver.class::isInstance)) {
+            DriverManager.registerDriver(new IngresDriver());
+        }
+    }
+}
