+++ plugin-jdbc-actianvector/src/main/java/io/kestra/plugin/jdbc/actianvector/package-info.java
@@ -1,0 +1,9 @@
+@PluginSubGroup(
+    description = "This sub-group of plugins contains tasks for accessing the Actian Vector database.",
+    categories = {
+        PluginSubGroup.PluginCategory.DATA
+    }
+)
+package io.kestra.plugin.jdbc.actianvector;
+
+import io.kestra.core.models.annotations.PluginSubGroup;
