+++ plugin-jdbc-actianvector/src/main/resources/metadata/actianvector.yaml
@@ -1,0 +1,8 @@
+group: io.kestra.plugin.jdbc.actianvector
+name: "actianvector"
+title: "Actian Vector"
+description: "Tasks that connect to Actian Vector via JDBC to run queries and loads."
+body: "Provide the Actian Vector JDBC URL, credentials, and SQL statements (query, batch, load, trigger) to read or write data; configure fetch type, fetch size, and output options to return rows or files from workflows."
+videos: []
+createdBy: "Kestra Core Team"
+managedBy: "Kestra Core Team"
