+++ plugin-jdbc-actianvector/src/test/resources/scripts/actianvector.sql
