--- plugin-jdbc-vectorwise/build.gradle
@@ -1,21 +1,0 @@
-project.description = 'Integrate Actian Vector (VectorWise) analytical databases with Kestra\'s JDBC plugin.'
-
-jar {
-    manifest {
-        attributes(
-                "X-Kestra-Name": project.name,
-                "X-Kestra-Title": "Actian Vector (VectorWise)",
-                "X-Kestra-Group": project.group + ".jdbc.vectorwise",
-                "X-Kestra-Description": project.description,
-                "X-Kestra-Version": project.version
-        )
-    }
-}
-
-dependencies {
-    jdbcDriver 'org.talend.libraries:iijdbc:6.0.0'
-    implementation project(':plugin-jdbc')
-
-    testImplementation project(':plugin-jdbc').sourceSets.test.output
-    testRuntimeOnly project(path: ':plugin-jdbc', configuration: 'shadow')
-}
