--- plugin-jdbc-vectorwise/src/main/java/io/kestra/plugin/jdbc/vectorwise/VetorwiseConnectionInterface.java
@@ -1,10 +1,0 @@
-package io.kestra.plugin.jdbc.vectorwise;
-
-import io.kestra.plugin.jdbc.JdbcConnectionInterface;
-
-public interface VetorwiseConnectionInterface extends JdbcConnectionInterface {
-    @Override
-    default String getScheme() {
-        return "jdbc:vectorwise";
-    }
-}
