--- plugin-jdbc-vectorwise/src/main/java/io/kestra/plugin/jdbc/vectorwise/package-info.java
@@ -1,9 +1,0 @@
-@PluginSubGroup(
-    description = "This sub-group of plugins contains tasks for accessing the Vectorwise database.",
-    categories = {
-        PluginSubGroup.PluginCategory.DATA
-    }
-)
-package io.kestra.plugin.jdbc.vectorwise;
-
-import io.kestra.core.models.annotations.PluginSubGroup;
