--- plugin-jdbc-vectorwise/src/test/resources/flows/vectorwise-listen.yml
@@ -1,18 +1,0 @@
-# not used as we cannot run it locally
-id: vectorwise-listen
-namespace: io.kestra.tests
-
-triggers:
-  - id: watch
-    type: io.kestra.plugin.jdbc.vectorwise.Trigger
-    sql: SELECT * FROM vectorwise_types
-    url: ???
-    username: ???
-    password: ???
-    fetchType: FETCH
-    interval: PT10S
-
-tasks:
-  - id: end
-    type: io.kestra.plugin.core.debug.Return
-    format: "{{task.id}} > {{taskrun.startDate}}"
