+++ settings.gradle
@@ -13,7 +13,7 @@ include 'plugin-jdbc-redshift'
 include 'plugin-jdbc-snowflake'
 include 'plugin-jdbc-sqlserver'
 include 'plugin-jdbc-trino'
-include 'plugin-jdbc-vectorwise'
+include 'plugin-jdbc-actianvector'
 include 'plugin-jdbc-vertica'
 include 'plugin-jdbc-dremio'
 include 'plugin-jdbc-arrow-flight'
